requirements.txt
matplotlib==3.8.0
numpy==2.2.4
pandas==2.2.3
pytest==7.4.0
python_dateutil==2.8.2
scikit_learn==1.1.3
seaborn==0.13.2
statsmodels==0.14.0
